Add account creation screen
Adds a new menu option for creating an email account, which takes the
user to a screen where they enter the account details (although it
doesn't actually do anything with them yet).

Updates the traps to use a $page var to keep track of where we are so
that we can SIGINT from any page that's not the main menu and return to
it, but still also SIGINT from the main menu itself to quit muttconf.

We can't just call `main` from the SIGINT trap because it will never
return, meaning subsequent SIGINTs are ignored. Instead we manually
return with a unique exit code which can be checked in the ERR trap.