Allow prompts to accept n-char inputs
Updates the prompt so that, when the 4th arg is `1`, input will be taken
immediately after the user enters a single character which is especially
useful in the main menu. It must be optional though or it breaks other
prompts used for things like email, names and other config options.