Add confirmation for account details
Adds support to `prompt` for passing in *all* available values to enable
confirmation prompts such as "Continue? [yN]" while still retaining the
option for a default if no value is passed.

This new prompt functionality is then used to confirm whether the
entered account details are correct after displaying them to the user.