Fix nested create_account calls
By calling create_account again with the given values when the user
needs to alter them, we could end up with a dozen nested function calls.

This all seems fine in theory, but the moment we do anything *after* the
confirmation, those actions will be repeated as many times as there were
calls to itself when the user needed to alter some details.

We fix that by moving the actual prompts to a new function, and looping
it as long as confirmation isn't available. This way, anything that
should only happen once can be done outside the loop and won't repeat.