IT'S A TRAP!
Finally fixes the traps so that they (seem to) work 100% as expected.

If you ran the script then hit 'n' to create an account, you could hit
^C to return to the main menu. BUT if you then hit 'n' again, followed
by another ^C, the SIGINT would be ignored and would be echoed to shell.

This commit makes it so the main menu is ran in a subshell so that we
can trap SIGINT in the child process instead and simply exit when ^C is
pressed from any page other than the main menu. If it IS the main menu,
we instead spit out a special exit code so that we can check that and
break out of the loop if it's a match when the subshell is exited.

Now you can go to the creation menu and hit ^C to return to the main
menu as many times as you want and it'll keep working. Success!

Source of inspiration: https://unix.stackexchange.com/a/67552/42182
(kill 0 isn't used here because it spits out "Terminated <PID>")