Save a basic single-account muttrc
Currently only has the settings required for built-in IMAP support, but
built-in SMTP and external IMAP/SMTP support can be added later.

The main thing is that it works, and it does. Woo!

Rather than using exec which can potentially be a bad idea (or simply
having the entire template inline in the script, which is ugly af), this
uses envsubst to substitute environment variables. We could manually
specify the vars to substitute, but it's easier to export in the subshell.

Also adds some rudimentary server/config detection based on domain. This
can eventually be improved by maybe doing a `dig MX <domain>` to help
detect Google Apps and other custom-domain email providers.