Load mailboxes to verify connection to IMAP server
Attempts to connect to the IMAP server using the given username (with
password from my mailpass script, which needs to be changed) to verify
that the settings are valid before saving them.

If curl returns 67 it means "Login denied", which translates to either
the username or password being wrong, or potentially "secure app access"
being disabled on Gmail/GSuite, so we prompt for that specially.

Seems to take a moment for secure app access to enable, so the user can
retry without having to go through the menus and enter everything again.
May take a couple tries, but eventually we get mailboxes which will be
used later to configure spoolfile, sent folder etc.