Give connection verification its own page
Verifying connections could end up going way off the screen with
multiple retries, so this puts it on its own page and also adds a
maximum number of attempts, which are displayed in the info text.

Once the retry limit is hit, we throw an error instead of prompting to
retry once again, followed by a "press any key" prompt which will take
the user back to the main menu where they can do whatever they want.