Prompt for (and save) IMAP password
LastPass CLI sucks so instead of instantly syncing and spitting the ID
out in response, we have to sync manually then parse it from lpass ls
and hope to god there isn't more than one entry.

After password is saved (or not in some cases), we pause for a second
before proceeding to the next screen so the user can read the message.